+++ backend/integrations/reddit/types.go
@@ -15,19 +15,19 @@ type SubReddit struct {
 
 // Post represents a Reddit post.
 type Post struct {
-	ID          string `json:"id"`
-	Title       string `json:"title"`
-	Author      string `json:"author"`
-	Score       int    `json:"score"`
-	Ups         int    `json:"ups"`   // Number of upvotes
-	Downs       int    `json:"downs"` // Number of downvotes (usually not directly exposed in v1 API)
-	URL         string `json:"url"`
-	Permalink   string `json:"permalink"`
-	CreatedAt   int64  `json:"created_utc"`
-	NumComments int    `json:"num_comments"`
-	Selftext    string `json:"selftext"`
-	IsSelf      bool   `json:"is_self"`
-	Subreddit   string `json:"subreddit"`
+	ID          string  `json:"id"`
+	Title       string  `json:"title"`
+	Author      string  `json:"author"`
+	Score       int     `json:"score"`
+	Ups         int     `json:"ups"`   // Number of upvotes
+	Downs       int     `json:"downs"` // Number of downvotes (usually not directly exposed in v1 API)
+	URL         string  `json:"url"`
+	Permalink   string  `json:"permalink"`
+	CreatedAt   float64 `json:"created_utc"`
+	NumComments int     `json:"num_comments"`
+	Selftext    string  `json:"selftext"`
+	IsSelf      bool    `json:"is_self"`
+	Subreddit   string  `json:"subreddit"`
 	AuthorInfo  *User
 	Comments    []*Comment
 	// Add other relevant fields from the post API response
@@ -35,27 +35,27 @@ type Post struct {
 
 // Comment represents a Reddit comment.
 type Comment struct {
-	ID        string `json:"id"`
-	Author    string `json:"author"`
-	Body      string `json:"body"`
-	Permalink string `json:"permalink"`
-	CreatedAt int64  `json:"created_utc"`
-	Score     int    `json:"score"`
-	Ups       int    `json:"ups"`   // Number of upvotes
-	Downs     int    `json:"downs"` // Number of downvotes (usually not directly exposed in v1 API)
-	ParentID  string `json:"parent_id"`
-	Depth     int    `json:"depth"`
+	ID        string  `json:"id"`
+	Author    string  `json:"author"`
+	Body      string  `json:"body"`
+	Permalink string  `json:"permalink"`
+	CreatedAt float64 `json:"created_utc"`
+	Score     int     `json:"score"`
+	Ups       int     `json:"ups"`   // Number of upvotes
+	Downs     int     `json:"downs"` // Number of downvotes (usually not directly exposed in v1 API)
+	ParentID  string  `json:"parent_id"`
+	Depth     int     `json:"depth"`
 	Comments  []*Comment
 	// Add other relevant comment fields
 }
 
 // User represents a Reddit user.
 type User struct {
-	ID               string `json:"id"`
-	Name             string `json:"name"`
-	Karma            int    `json:"total_karma"`
-	CreatedAt        int64  `json:"created_utc"`
-	IsGold           bool   `json:"is_gold"`
-	HasVerifiedEmail bool   `json:"has_verified_email"`
+	ID               string  `json:"id"`
+	Name             string  `json:"name"`
+	Karma            int     `json:"total_karma"`
+	CreatedAt        float64 `json:"created_utc"`
+	IsGold           bool    `json:"is_gold"`
+	HasVerifiedEmail bool    `json:"has_verified_email"`
 	// Add other relevant user-related fields
 }
